Create pl.json
Line 27: I've abbreviated all the seconds to "s" because a) I wasn't sure what the difference between the first second and the second second was (I assumed one of those is for fractions but still, not sure), and b) the declension of "second" in Polish is slightly problematic, the plural for 2-4 seconds is different than the plural for 5+ seconds, so I'll leave it as it is for now.
Line 261 & 262: I've changed the date format do DD/MM/YY, as used in Poland.
Misc:
I've opted against translating feature names, as they seem to be embedded in the graphics on the main page (and they're not localised in the French version either), if you'd like them localised, feel free to let me know.
Whenever another user is to be mentioned I inserted the word "user" before that to avoid grammatical gender mismatches, which makes certain lines a bit longer than they were (cf. line 247 & 248, for instance).
Lines 285 and below are in French, so I wasn't sure if they should be translated as well. If so, please let me know, I think I know French well enough to manage these (also, my fiancée knows French way better than I do, so she could help me in any case).